backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlmodel
passlib
bcrypt==4.0.1
python-jose[cryptography]
opencv-python-headless
pytesseract
Pillow
python-multipart
pdf2image
google-generativeai
python-dotenv
